requirements.txt
# Runtime dependencies
python-telegram-bot>=20,<22
discord.py>=2.3,<3
python-dotenv>=1.0,<2
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
motor>=3.4.0
pymongo>=4.6.0
requests
